+++ module/actor/actor-sheet.js
@@ -8,6 +8,9 @@ import {
 } from "./databar.js";
 import { buildCrewOptionsArray } from "./crew.js";
 
+const { ActorSheet } = foundry.appv1.sheets;
+const { Dialog } = foundry.appv1.api;
+
 /**
  * Extend the basic ActorSheet for a basic Coriolis character
  * @extends {ActorSheet}
@@ -35,18 +38,6 @@ export class yzecoriolisActorSheet extends ActorSheet {
       ],
     });
   }
-
-  /**
-   * Activate a named TinyMCE text editor
-   * @param {string} name             The named data field which the editor modifies.
-   * @param {object} options          TinyMCE initialization options passed to TextEditor.create
-   * @param {string} initialContent   Initial text content for the editor area.
-   */
-  activateEditor(name, options = {}, initialContent = "") {
-    const customOptions = { ...options, body_class: "charnotes-edit-body" };
-    super.activateEditor(name, customOptions, initialContent);
-  }
-  /* -------------------------------------------- */
 
   /** @override */
   async getData(options) {