+++ module/actor/ship-sheet.js
@@ -21,6 +21,8 @@ import {
 import { toggleShipModule } from "../item/ship-module.js";
 import { CoriolisModifierDialog } from "../coriolisRollModifier.js";
 
+const { ActorSheet } = foundry.appv1.sheets;
+
 /**
  * Extend the basic ActorSheet for a basic Coriolis ship sheet
  * @extends {ActorSheet}
@@ -50,17 +52,6 @@ export class yzecoriolisShipSheet extends ActorSheet {
       ],
     });
   }
-  /**
-   * Activate a named TinyMCE text editor
-   * @param {string} name             The named data field which the editor modifies.
-   * @param {object} options          TinyMCE initialization options passed to TextEditor.create
-   * @param {string} initialContent   Initial text content for the editor area.
-   */
-  activateEditor(name, options = {}, initialContent = "") {
-    const customOptions = { ...options, body_class: "charnotes-edit-body" };
-    super.activateEditor(name, customOptions, initialContent);
-  }
-
   async getData(options) {
     const baseData = super.getData(options);
     let stats = {};