+++ module/darkness-points.js
@@ -5,6 +5,9 @@
 // model in future foundry allows for custom world setting permission, this code
 // could be simplified.
 
+const { Application } = foundry.appv1.api;
+const { renderTemplate } = foundry.applications.handlebars;
+
 export async function addDarknessPoints(points) {
   // local user cache
   let dpObj = getDarknessPointsForUserID(game.user.id);
@@ -76,9 +79,9 @@ async function showDarknessPoints(totalPoints) {
   let gmUser = gmList[0];
 
   let messageData = {
-    user: gmUser.id,
+    author: gmUser.id,
     speaker: ChatMessage.getSpeaker({ user: gmUser }),
-    whisper: gmList,
+    whisper: gmList.map((user) => user.id),
   };
 
   const dpData = {
@@ -107,7 +110,7 @@ export class DarknessPointDisplay extends Application {
   }
 
   static get defaultOptions() {
-    return mergeObject(super.defaultOptions, {
+    return foundry.utils.mergeObject(super.defaultOptions, {
       id: "coriolis-darness-points-display",
       template:
         "systems/yzecoriolis/templates/darkness-points/darkness-points-display.html",